+++ CryptoBank/Models/DepositWithdrawalInfo.cs
@@ -6,7 +6,7 @@ using System.Web;
 namespace CryptoBank.Models {
     public class DepositWithdrawalInfo {
         public string IBAN { get; set; }
-        public decimal amount { get; set; }
+        public double amount { get; set; }
         public string currency { get; set; }
         public bool deposit { get; set; }
     }
